+++ spec/directions/composite_page_container/main_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-RSpec.describe Kitchen::Directions::EocCompositePageContainer do
+RSpec.describe Kitchen::Directions::CompositePageContainer do
   it 'calls v1' do
-    expect_any_instance_of(Kitchen::Directions::EocCompositePageContainer::V1).to receive(:bake)
+    expect_any_instance_of(Kitchen::Directions::CompositePageContainer::V1).to receive(:bake)
       .with(container_key: 'some-section', uuid_key: 'some-section', metadata_source: 'metadata', content: 'content', append_to: 'append')
     described_class.v1(container_key: 'some-section', uuid_key: 'some-section', metadata_source: 'metadata', content: 'content', append_to: 'append')
   end